==> 20.Creating_new_databases/3.Creating_new_db.ipynb <==
# Needed modules

# Load needed libraries
import pandas as pd
import os

# https://kioku-space.com/en/jupyter-skip-execution/
from IPython.core.magic import register_cell_magic # type: ignore

@register_cell_magic
def skip(line, cell):
    return

# Pickle save

%%skip
# =============================================================================
# Save the variables
# =============================================================================
variables_dict = {
}

# =============================================================================
# main function
# =============================================================================
def data_save_load(option, dict_variables=None):
    """
    This function is used to save or load data for the jupyter notebook
    """
    path_folder = "ipynb_db"  # Folder to save variables
    os.makedirs(path_folder, exist_ok=True)  # Create folder if not exist
    path_file = os.path.join(path_folder, "variables.pkl") # Path to save the variables

    if option == "save":
        with open(path_file, "wb") as f:
            pickle.dump(dict_variables, f)
    elif option == "load":
        with open(path_file, "rb") as f:
            variables = pickle.load(f)
        # Now load the variables
        for key, value in variables.items():
            variables[key] = value

# =============================================================================
# Call the function
# =============================================================================
data_save_load(option="save",
               dict_variables=variables_dict)

# 1. Load data

# Load new negative data:
neg_data = pd.read_csv("./2.Hallmarks_vs_NegData/negative_data_after_recaught.csv", sep=",", header=0)
neg_data.sort_values(["sseqid", "sstart"])
print(neg_data.shape)
print(neg_data.dtypes)
neg_data.head()

# Load neg recaught data, i.e., data to add to the positive elements
neg_recaught_data = pd.read_csv("./2.Hallmarks_vs_NegData/neg_data_recaught_hits.csv", sep=",", header=0)
print(neg_recaught_data.shape)
print(neg_recaught_data.dtypes)
neg_recaught_data.head()

# Load positive data:
pos_data = pd.read_csv("./1.new_data/positive_database.csv", sep=",", header=0)
print(pos_data.shape)
print(pos_data.dtypes)
pos_data.head()

# 2. Add recaught data to the positive data

# Add neg_recaught_data to pos_data
pos_data = pd.concat([pos_data, neg_recaught_data], axis=0)

# Sort by ssqid, sstart
pos_data.sort_values(by=["sseqid", "sstart"], inplace=True, ignore_index=True)

# Check data
print(pos_data.shape)
print(pos_data.dtypes)
pos_data.head()

# 3. Check if the negative and positive data for included or overlaps

# Copy og data
pos_data_check = pos_data.copy()
neg_data_check = neg_data.copy()

# Create interval columns
pos_data_check["interval"] = pos_data_check.apply(lambda x: pd.Interval(left=x["sstart"], right=x["send"], closed="both"), axis=1)
neg_data_check["interval"] = neg_data_check.apply(lambda x: pd.Interval(left=x["sstart"], right=x["send"], closed="both"), axis=1)

print(pos_data_check.loc[:, ["sseqid", "sstart", "send", "interval"]].head())
print(neg_data_check.loc[:, ["sseqid", "sstart", "send", "interval"]].head())

# Group the elements
pos_data_grouped = pos_data_check.groupby("sseqid")
neg_data_grouped = neg_data_check.groupby("sseqid")

# Let's check for each group of positive elements if there is any negative if the elements are exactly the same, overlap or are inside one another
same_elems = []
overlaping_elems = []
inside_elems = []
inside_elems2 = []
for name1, data1 in pos_data_grouped:
    for name2, data2 in neg_data_grouped:
        if name2 != name1:
            # print(f">> SKIPPING Chrom. Chromosome in negative elements {name2} is different from the one in positive elements {name1}")
            continue
        else:
            print("="*50)
            print(f"Checking chromosome {name1}")
            # Checking for same coordinates
            for _, elem1 in data1.iterrows():
                for _, elem2 in data2.iterrows():
                    if elem1["interval"] == elem2["interval"]:
                        print(f"Found same element in chromosome {name1}")
                        same_elems.append(elem1)
                    elif elem1["interval"].overlaps(elem2["interval"]):
                        print(f"Found overlaping element in chromosome {name1}")
                        overlaping_elems.append([elem1, elem2])
                    else:
                        if (elem1["sstart"] > elem2["sstart"] and elem1["sstart"] < elem2["send"] /
                            elem1["send"] < elem2["send"] and elem1["send"] > elem2["sstart"]):
                            print(f"Found element inside another element in chromosome {name1}")
                            inside_elems.append(elem1)
                        if (elem2["sstart"] > elem1["sstart"] and elem2["sstart"] < elem1["send"] /
                            elem2["send"] < elem1["send"] and elem2["send"] > elem1["sstart"]):
                            print(f"Found element inside another element in chromosome {name1}")
                            inside_elems2.append(elem1)

print("*"*50)
print(f"Found {len(same_elems)} elements that are the same")
print(f"Found {len(overlaping_elems)} elements that are overlaping")
print(f"Found {len(inside_elems)} elements that are inside another element")
print(f"Found {len(inside_elems2)} elements that are inside another element")


# 4. Save data

# Prepare paths
path_folder = "./3.new_data_v2"
path_pos_data = os.path.join(path_folder, "positive_database.csv")
path_neg_data = os.path.join(path_folder, "negative_database.csv")

# save data
os.makedirs(path_folder, exist_ok=True)
pos_data.to_csv(path_pos_data, sep=",", header=True, index=False)
neg_data.to_csv(path_neg_data, sep=",", header=True, index=False)
