==> 20.Creating_new_databases/4.Neg_data_vs_CDS.ipynb <==
# Code

## Modules

import pandas as pd
import subprocess
from Bio import SeqIO
from Bio.Seq import Seq
from Bio.SeqRecord import SeqRecord

## Import data

### Negative data base

# Read negative data
neg_data = pd.read_csv('./3.new_data_v2/negative_database.csv', sep=',', header=0)
print(neg_data.shape)
print(neg_data.dtypes)
neg_data.head()

# Get what's needed from the negative data
neg_data_bed = neg_data[["sseqid", "sstart", "send"]].copy()
print(neg_data_bed.head())

# Get the type of data
neg_data_bed.dtypes

### CDS data base

path_cds = './4.Neg_data_vs_CDS/data/TriTrypDB-68_LinfantumJPCM5_AnnotatedCDSs.fasta'
cds_data = []
with open(path_cds, "r") as handle:
    for record in SeqIO.parse(handle, "fasta"):
        # print(f"{record.description}")  # Here are all the data needed
        start, end = record.description.split("|")[3].split(":")[1].replace("(-)", "").replace("(+)","").strip().split("-")  # start and end coordinates
        sense = record.description.split("|")[3].split(":")[1].split("(")[1].replace(")","").strip()  # sense
        chr_num = record.description.split("|")[3].split(":")[0].split("=")[1]  # chromosome number
        attribute = record.description.split("|")[2].strip()
        # Now let's create a bed file with chr_num, start and end only, with tab sep
        cds_data.append([chr_num, start, end, sense, attribute])
        # print(f"Chromosome: {chr_num}, Start: {start}, End: {end}, Attribute: {attribute}")
print(len(cds_data))
[print(x) for x in cds_data[:5]]

# Transform the list into a dataframe
cds_data = pd.DataFrame(cds_data, columns=["sseqid", "sstart", "send", "sense", "attribute"])
cds_data[['sstart', 'send']] = cds_data[['sstart', 'send']].apply(pd.to_numeric)
print(cds_data.dtypes)
print(cds_data.head())

## Compare `neg_data_bed` against `cds_data`

## And check with our data have CDS inside
neg_with_cds = {}
for index, row in neg_data_bed.iterrows():
    overlapping_cds = cds_data[
       (
           (
               ((cds_data['sstart'] >= row['sstart']) & (cds_data['sstart'] <= row['send'])) |
               ((cds_data['send'] <= row['send']) & (cds_data['send'] >= row['sstart']))
            ) |  # Now let's do for the "neg_data" that are inside the CDS
           (
               (cds_data['sstart'] <= row['sstart']) & (cds_data['send'] >= row['send'])
            )
        ) &
       (cds_data['sseqid'] == row['sseqid'])
        ]
    if not overlapping_cds.empty:
        neg_with_cds[f'{row['sseqid']}_{row['sstart']}_{row['send']}'] = overlapping_cds
    else:
        continue

for key, value in neg_with_cds.items():
    print(f"Key: {key}")
    if len(value) >= 1:
        for index, row in value.iterrows():
            print(f"\tValue: {list(row)}")

print(f'Therea are {len(neg_with_cds)}/{len(neg_data)} elements in the negative data that overlap with CDS')
print(f'There are {len(neg_data) - len(neg_with_cds)} elements in the negative data that do not overlap with CDS')

count_cds = 0
for key, value in neg_with_cds.items():
    count_cds += value.shape[0]
print(f'There are {count_cds}/{len(cds_data)} CDS that overlap with the negative data')
print(f'There are {len(cds_data) - count_cds} CDS that do not overlap with the negative data')


## Export as CSV

### Export the relationship with CDS

pre_csv = []
for key, value in neg_with_cds.items():
    for index, row in value.iterrows():
        pre_csv.append([key, f'{row['sseqid']}|{row['sstart']}|{row['send']}|{row['sense']}|{row['attribute']}'])
print(len(pre_csv))
[print(x) for x in pre_csv[:5]]

pre_csv = pd.DataFrame(pre_csv, columns=["neg_key", "cds_key"])

# Save the data
pre_csv.to_csv('./4.Neg_data_vs_CDS/new_data/neg_with_cds.csv', sep=',', header=True, index=False)

### Export the onlye the rows in neg data that solap with CDS

# Get the elemens that overlap with CDS
neg_data_only_overlaps= []
for key, value in neg_with_cds.items():
    key_chr, key_start, key_end = key.split("_")[0], key.split("_")[1], key.split("_")[2]
    neg_data_only_overlaps.append([key_chr, key_start, key_end])

neg_data_only_overlaps = pd.DataFrame(neg_data_only_overlaps, columns=["sseqid", "sstart", "send"])
neg_data_only_overlaps[['sstart', 'send']] = neg_data_only_overlaps[['sstart', 'send']].apply(pd.to_numeric)
print(neg_data_only_overlaps.dtypes)
print(len(neg_data_only_overlaps))

# Now extract the from neg data the values that are in neg_data_only_solaps and the ones that are now.
new_neg_data_overlaps_with_cds = neg_data.merge(neg_data_only_overlaps, on=["sseqid", "sstart", "send"], how="inner")
new_neg_data_no_overlaps_with_cds = neg_data.merge(neg_data_only_overlaps, on=["sseqid", "sstart", "send"], how="outer", indicator=True).query('_merge == "left_only"').drop('_merge', axis=1)

# Print the data len
print(len(new_neg_data_overlaps_with_cds))
print(len(new_neg_data_no_overlaps_with_cds))

# Save both to CSV
new_neg_data_overlaps_with_cds.to_csv('./4.Neg_data_vs_CDS/new_data/negative_database_match_CDS.csv', sep=',', header=True, index=False)
new_neg_data_no_overlaps_with_cds.to_csv('./4.Neg_data_vs_CDS/new_data/negative_database_nomatch.csv', sep=',', header=True, index=False)

## Create a GFF file

### Neg data that overlaps with CDS

gff_neg_with_cds = []
column_names = ['seqname', 'source', 'feature', 'start', 'end', 'score', 'strand', 'frame', 'attribute']
for key, value in neg_with_cds.items():
    key_chr, key_start, key_end = key.split("_")[0], key.split("_")[1], key.split("_")[2]
    attr_str = ""
    for _, row in value.iterrows():
        attr_str += f"{row['sseqid']}|{row['sstart']}|{row['send']}|{row['sense']}|{row['attribute']};"
    gff_neg_with_cds.append(
        {
            'seqname': key_chr,
            'source': 'CBM',
            'feature': 'Negative_Repetead_with_CDS',
            'start': key_start,
            'end': key_end,
            'score': '.',
            'strand': '+',
            'frame': '.',
            'attribute': f'Related_to={attr_str}'
        }
    )
print(len(gff_neg_with_cds))
[print(x) for x in gff_neg_with_cds[:5]]

gff_neg_with_cds = pd.DataFrame(gff_neg_with_cds, columns=column_names)

gff_neg_with_cds.to_csv('./4.Neg_data_vs_CDS/new_data/neg_with_cds.gff', sep='\t', header=False, index=False)

#### Neg data that does not overlap with CDS

gff_neg_no_cds = []
column_names = ['seqname', 'source', 'feature', 'start', 'end', 'score', 'strand', 'frame', 'attribute']
for index, row in new_neg_data_no_overlaps_with_cds.iterrows():
    gff_neg_no_cds.append(
        {
            'seqname': row['sseqid'],
            'source': 'CBM',
            'feature': 'Negative_Repetead_no_CDS',
            'start': row['sstart'],
            'end': row['send'],
            'score': '.',
            'strand': '+',
            'frame': '.',
            'attribute': '.'
        }
    )
print(len(gff_neg_no_cds))
[print(x) for x in gff_neg_no_cds[:5]]

gff_neg_no_cds = pd.DataFrame(gff_neg_no_cds, columns=column_names)
gff_neg_no_cds.to_csv('./4.Neg_data_vs_CDS/new_data/neg_no_cds.gff', sep='\t', header=False, index=False)
